requirements.txt
django==5.0
celery==5.3.6
redis==5.0.1
flower==2.0.1
psycopg2-binary==2.9.9
watchfiles==0.21.0
requests==2.31.0
daphne==4.0.0
channels==4.0.0
channels-redis==4.1.0
git+https://github.com/celery/django-celery-beat.git#egg=django-celery-beat  # update when new version supports Django 5
factory-boy==3.3.0
coverage==7.4.0
pytest==7.4.4
pytest-django==4.7.0
pytest-factoryboy==2.6.0
pytest-cov==4.1.0
Pillow==10.1.0
gunicorn==21.2.0
uvicorn[standard]==0.25.0
